Test range

See merge request csce-3600/fall-2024/section-004/major-1/group-10/major1!2

+++ major1.c
@@ -1,9 +1,8 @@
 #include "major1.h"
 
-
 int main(){
     //Operand
-    unsigned int num;
+    long long int num;
     while(1==1){
         //Display menu
         printf("Enter the menu option for the operation to perform: \n");
@@ -17,18 +16,18 @@ int main(){
         //Menu selection
         int choice;
         int validNum = 0;
-        scanf("%d", &choice);
+        scanf("%u", &choice);
         switch(choice){
             case 1:
                 while(validNum == 0){
                     //Prompt for integer between 1 and 4294967295, inclusively
                     printf("Enter a 32-bit number (>= 1 and <= 4294967295, inclusively): ");
-                    scanf("%u", &num);
+                    scanf("%lld", &num);
                     if(num >= 1 && num <= 4294967295){
                         //Count Leading Zeroes
                         validNum = 1;
-                        int leading0s = CountLeadingZeroes(num);
-                        printf("\e[1mThe number of leading zeroes in %u is %d\n\e[m", num, leading0s);
+                        int leading0s = CountLeadingZeroes((unsigned int)num);
+                        printf("\e[1mThe number of leading zeroes in %u is %d\n\e[m", (unsigned int)num, leading0s);
                     }
                 }
                     break;
@@ -37,12 +36,12 @@ int main(){
                 while(validNum == 0){
                     // Prompt for a 32-bit number
                     printf("Enter a 32-bit number (>= 1 and <= 4294967295, inclusively): ");
-                    scanf("%u", &num);
+                    scanf("%lld", &num);
                     if(num >= 1 && num <= 4294967295){
                         // Perform endian swap
                         validNum = 1;
                         unsigned int endSwapped = endianSwap((unsigned int)num);
-                        printf("\e[1mEndian swap of %u gives %u\n\e[m", num, endSwapped);
+                        printf("\e[1mEndian swap of %u gives %u\n\e[m",(unsigned int) num, endSwapped);
                     }
                 }
                 break;
@@ -51,7 +50,7 @@ int main(){
 		while(validNum == 0){
 			//Prompt for a 32-bit number and number of shift positions
 			printf("Enter a 32-bit number (>= 1 and <= 4294967295, inclusively): ");
-			scanf("%u", &num);
+			scanf("%lld", &num);
 			unsigned int pos;// Used to store the number of bit positions to be shifted
 			printf("Enter number of positions to rotate-right the input (between 0 and 31, inclusively): ");
 			scanf("%u", &pos);
@@ -59,7 +58,7 @@ int main(){
 				// Perform bit rotation
 				validNum = 1;
 				unsigned int rotate = RotateInt((unsigned int) num, (unsigned int) pos); // Call the rotate function
-				printf("\e[1m%u rotated by %u positions gives: %u\n\e[m", num, pos, rotate);// Outputs: Rotated num
+				printf("\e[1m%u rotated by %u positions gives: %u\n\e[m", (unsigned int) num, pos, rotate);// Outputs: Rotated num
 			}
 		}
                 break;
@@ -68,12 +67,12 @@ int main(){
                 while(validNum == 0){
                     // Prompt for a 32-bit number
                     printf("Enter a 32-bit number (>= 1 and <= 4294967295, inclusively): ");
-                    scanf("%u", &num);
+                    scanf("%lld", &num);
                     if(num >= 1 && num <= 4294967295){
                         // Compute parity
                         validNum = 1;
                         int parity = compute_parity((unsigned int)num); // Call the function from parity.c
-                        printf("\e[1mParity of %u is %d\n\e[m", num, parity);  // Output result: 0 for even, 1 for odd
+                        printf("\e[1mParity of %u is %d\n\e[m", (unsigned int)num, parity);  // Output result: 0 for even, 1 for odd
                     }
                 }
                 break;
@@ -86,3 +85,4 @@ int main(){
     }
     return 0;
 }
+
